[refactor] msw 바인딩 data 분리
