+++ src/mocks/data/studyListMockData.ts
