+++ src/mocks/data/userMockData.ts
@@ -1,0 +1,439 @@
+import { AvgStats, CompletedStudyItem, OnGoingStudyItem, UpcomingStudyItem, UserProfile } from '../../types/interface';
+
+export const upcomingStudies: UpcomingStudyItem[] = [
+  {
+    recruit_id: 1,
+    gatherDate: '2025.03.01 ~ 2025.08.01',
+    title: '9급 공무원 준비',
+    enterPoint: 3000,
+    tag: '공시',
+    category: '공무원',
+    weeklyStudyTime: 10,
+  },
+  {
+    recruit_id: 2,
+    gatherDate: '2025.04.10 ~ 2025.09.10',
+    title: '토익 900 목표반',
+    enterPoint: 2500,
+    tag: '토익',
+    category: '시험',
+    weeklyStudyTime: 8,
+  },
+  {
+    recruit_id: 3,
+    gatherDate: '2025.05.15 ~ 2025.11.15',
+    title: '회계 자격증반',
+    enterPoint: 2000,
+    tag: '회계',
+    category: '시험',
+    weeklyStudyTime: 8,
+  },
+  {
+    recruit_id: 4,
+    gatherDate: '2025.06.01 ~ 2025.12.01',
+    title: '한국사 1급 도전',
+    enterPoint: 1500,
+    tag: '한국사',
+    category: '시험',
+    weeklyStudyTime: 6,
+  },
+  {
+    recruit_id: 5,
+    gatherDate: '2025.07.01 ~ 2025.12.31',
+    title: '공인중개사 대비',
+    enterPoint: 1800,
+    tag: '공인중개사',
+    category: '시험',
+    weeklyStudyTime: 7,
+  },
+  {
+    recruit_id: 6,
+    gatherDate: '2025.08.01 ~ 2026.02.01',
+    title: '컴활 1급 단기완성',
+    enterPoint: 2200,
+    tag: '컴활',
+    category: '시험',
+    weeklyStudyTime: 9,
+  },
+  {
+    recruit_id: 7,
+    gatherDate: '2025.09.15 ~ 2026.03.15',
+    title: 'JLPT N1 스터디',
+    enterPoint: 1700,
+    tag: 'JLPT',
+    category: '시험',
+    weeklyStudyTime: 6,
+  },
+  {
+    recruit_id: 8,
+    gatherDate: '2025.10.10 ~ 2026.04.10',
+    title: '행정법 기출 분석',
+    enterPoint: 2000,
+    tag: '공시',
+    category: '공무원',
+    weeklyStudyTime: 8,
+  },
+  {
+    recruit_id: 9,
+    gatherDate: '2025.11.05 ~ 2026.05.05',
+    title: '독서 & 글쓰기',
+    enterPoint: 2500,
+    tag: '자기계발',
+    category: '기타',
+    weeklyStudyTime: 10,
+  },
+  {
+    recruit_id: 10,
+    gatherDate: '2025.12.01 ~ 2026.06.01',
+    title: '경제학 기본개념',
+    enterPoint: 3000,
+    tag: '경제학',
+    category: '기타',
+    weeklyStudyTime: 5,
+  },
+  {
+    recruit_id: 11,
+    gatherDate: '2026.01.15 ~ 2026.07.15',
+    title: '한국어능력시험',
+    enterPoint: 1800,
+    tag: '한국어',
+    category: '시험',
+    weeklyStudyTime: 7,
+  },
+  {
+    recruit_id: 12,
+    gatherDate: '2026.02.20 ~ 2026.08.20',
+    title: '법원직 시험 준비',
+    enterPoint: 2200,
+    tag: '공시',
+    category: '공무원',
+    weeklyStudyTime: 9,
+  },
+  {
+    recruit_id: 13,
+    gatherDate: '2026.03.10 ~ 2026.09.10',
+    title: '독일어 B2 목표',
+    enterPoint: 2600,
+    tag: '독어',
+    category: '어학',
+    weeklyStudyTime: 11,
+  },
+  {
+    recruit_id: 14,
+    gatherDate: '2026.04.05 ~ 2026.10.05',
+    title: '사회복지사 1급',
+    enterPoint: 1900,
+    tag: '사회복지사',
+    category: '시험',
+    weeklyStudyTime: 8,
+  },
+  {
+    recruit_id: 15,
+    gatherDate: '2026.05.01 ~ 2026.11.01',
+    title: '공기업 NCS 대비',
+    enterPoint: 1300,
+    tag: '공기업',
+    category: '취업',
+    weeklyStudyTime: 4,
+  },
+];
+
+export const ongoingStudies: OnGoingStudyItem[] = [
+  {
+    registerId: 1,
+    gatherDate: '2024.11.01 ~ 2025.05.01',
+    title: 'SAT 준비 스터디',
+    enterPoint: 5000,
+    tag: 'SAT',
+    category: '시험',
+    weeklyStudyTime: 12,
+  },
+  {
+    registerId: 2,
+    gatherDate: '2024.10.01 ~ 2025.04.01',
+    title: '영어 회화 스터디',
+    enterPoint: 3000,
+    tag: '영어',
+    category: '어학',
+    weeklyStudyTime: 6,
+  },
+  {
+    registerId: 3,
+    gatherDate: '2024.09.15 ~ 2025.03.15',
+    title: '공무원 행정법',
+    enterPoint: 4000,
+    tag: '공시',
+    category: '공무원',
+    weeklyStudyTime: 10,
+  },
+  {
+    registerId: 4,
+    gatherDate: '2024.12.01 ~ 2025.06.01',
+    title: '토익 스피킹반',
+    enterPoint: 3500,
+    tag: '토익',
+    category: '어학',
+    weeklyStudyTime: 8,
+  },
+  {
+    registerId: 5,
+    gatherDate: '2024.08.20 ~ 2025.02.20',
+    title: '데이터 분석 스터디',
+    enterPoint: 4500,
+    tag: '데이터분석',
+    category: '코딩',
+    weeklyStudyTime: 9,
+  },
+];
+
+export const completedStudies: CompletedStudyItem[] = [
+  {
+    studyId: 1,
+    deductedPoint: 100,
+    obtainedPoint: 200,
+    gatherDate: '2024.12.15 ~ 2025.01.15',
+    title: 'C++ 프로그래밍',
+    enterPoint: 5000,
+    tag: 'C++',
+    category: '코딩',
+    weeklyStudyTime: 8,
+  },
+  {
+    studyId: 2,
+    deductedPoint: 50,
+    obtainedPoint: 1500,
+    gatherDate: '2024.11.10 ~ 2024.12.10',
+    title: '웹 개발 프로젝트',
+    enterPoint: 3000,
+    tag: '웹 개발',
+    category: '코딩',
+    weeklyStudyTime: 10,
+  },
+  {
+    studyId: 3,
+    deductedPoint: 2000,
+    obtainedPoint: 350,
+    gatherDate: '2024.10.05 ~ 2024.11.05',
+    title: 'Java 알고리즘',
+    enterPoint: 4000,
+    tag: 'Java',
+    category: '코딩',
+    weeklyStudyTime: 12,
+  },
+  {
+    studyId: 4,
+    deductedPoint: 120,
+    obtainedPoint: 300,
+    gatherDate: '2024.09.20 ~ 2024.10.20',
+    title: '공무원 행정법',
+    enterPoint: 2500,
+    tag: '공시',
+    category: '공무원',
+    weeklyStudyTime: 9,
+  },
+  {
+    studyId: 5,
+    deductedPoint: 90,
+    obtainedPoint: 220,
+    gatherDate: '2024.08.15 ~ 2024.09.15',
+    title: '토익 완성반',
+    enterPoint: 3200,
+    tag: '토익',
+    category: '어학',
+    weeklyStudyTime: 7,
+  },
+  {
+    studyId: 6,
+    deductedPoint: 250,
+    obtainedPoint: 500,
+    gatherDate: '2024.07.10 ~ 2024.08.10',
+    title: '데이터 분석 스터디',
+    enterPoint: 4500,
+    tag: '데이터분석',
+    category: '코딩',
+    weeklyStudyTime: 10,
+  },
+  {
+    studyId: 7,
+    deductedPoint: 300,
+    obtainedPoint: 600,
+    gatherDate: '2024.06.05 ~ 2024.07.05',
+    title: 'Python 입문',
+    enterPoint: 4000,
+    tag: 'Python',
+    category: '코딩',
+    weeklyStudyTime: 8,
+  },
+  {
+    studyId: 8,
+    deductedPoint: 400,
+    obtainedPoint: 800,
+    gatherDate: '2024.05.01 ~ 2024.06.01',
+    title: '면접 대비반',
+    enterPoint: 3800,
+    tag: '면접',
+    category: '취업',
+    weeklyStudyTime: 6,
+  },
+  {
+    studyId: 9,
+    deductedPoint: 150,
+    obtainedPoint: 320,
+    gatherDate: '2024.04.10 ~ 2024.05.10',
+    title: '컴퓨터 네트워크',
+    enterPoint: 4200,
+    tag: '네트워크',
+    category: '기타',
+    weeklyStudyTime: 9,
+  },
+  {
+    studyId: 10,
+    deductedPoint: 180,
+    obtainedPoint: 370,
+    gatherDate: '2024.03.05 ~ 2024.04.05',
+    title: '시사 토론반',
+    enterPoint: 2800,
+    tag: '토론',
+    category: '기타',
+    weeklyStudyTime: 7,
+  },
+  {
+    studyId: 11,
+    deductedPoint: 60,
+    obtainedPoint: 150,
+    gatherDate: '2024.02.01 ~ 2024.03.01',
+    title: '공인중개사 대비',
+    enterPoint: 3500,
+    tag: '공인중개사',
+    category: '시험',
+    weeklyStudyTime: 10,
+  },
+  {
+    studyId: 12,
+    deductedPoint: 220,
+    obtainedPoint: 400,
+    gatherDate: '2024.01.10 ~ 2024.02.10',
+    title: '기초 통계학',
+    enterPoint: 3900,
+    tag: '통계',
+    category: '기타',
+    weeklyStudyTime: 8,
+  },
+  {
+    studyId: 13,
+    deductedPoint: 80,
+    obtainedPoint: 180,
+    gatherDate: '2023.12.05 ~ 2024.01.05',
+    title: '독서 토론 스터디',
+    enterPoint: 2600,
+    tag: '독서',
+    category: '기타',
+    weeklyStudyTime: 6,
+  },
+  {
+    studyId: 14,
+    deductedPoint: 90,
+    obtainedPoint: 200,
+    gatherDate: '2023.11.01 ~ 2023.12.01',
+    title: '일본어 초급반',
+    enterPoint: 3300,
+    tag: '일본어',
+    category: '어학',
+    weeklyStudyTime: 7,
+  },
+  {
+    studyId: 15,
+    deductedPoint: 250,
+    obtainedPoint: 500,
+    gatherDate: '2023.10.10 ~ 2023.11.10',
+    title: 'SQL 마스터 과정',
+    enterPoint: 4100,
+    tag: '데이터베이스',
+    category: '코딩',
+    weeklyStudyTime: 10,
+  },
+  {
+    studyId: 16,
+    deductedPoint: 200,
+    obtainedPoint: 400,
+    gatherDate: '2023.09.05 ~ 2023.10.05',
+    title: 'AI 기초 스터디',
+    enterPoint: 4300,
+    tag: '인공지능',
+    category: '코딩',
+    weeklyStudyTime: 9,
+  },
+  {
+    studyId: 17,
+    deductedPoint: 140,
+    obtainedPoint: 300,
+    gatherDate: '2023.08.01 ~ 2023.09.01',
+    title: '주식 투자 스터디',
+    enterPoint: 3700,
+    tag: '금융',
+    category: '기타',
+    weeklyStudyTime: 8,
+  },
+  {
+    studyId: 18,
+    deductedPoint: 120,
+    obtainedPoint: 250,
+    gatherDate: '2023.07.10 ~ 2023.08.10',
+    title: '여행 영어 회화',
+    enterPoint: 2900,
+    tag: '영어',
+    category: '어학',
+    weeklyStudyTime: 6,
+  },
+  {
+    studyId: 19,
+    deductedPoint: 300,
+    obtainedPoint: 600,
+    gatherDate: '2023.06.05 ~ 2023.07.05',
+    title: '공기업 NCS 대비',
+    enterPoint: 4500,
+    tag: '공기업',
+    category: '시험',
+    weeklyStudyTime: 10,
+  },
+  {
+    studyId: 20,
+    deductedPoint: 110,
+    obtainedPoint: 240,
+    gatherDate: '2023.05.01 ~ 2023.06.01',
+    title: 'GRE 수학 스터디',
+    enterPoint: 3500,
+    tag: 'GRE',
+    category: '시험',
+    weeklyStudyTime: 7,
+  },
+  {
+    studyId: 21,
+    deductedPoint: 130,
+    obtainedPoint: 280,
+    gatherDate: '2023.04.10 ~ 2023.05.10',
+    title: '국제정치학 스터디',
+    enterPoint: 3400,
+    tag: '정치',
+    category: '기타',
+    weeklyStudyTime: 8,
+  },
+];
+
+export const profileData: UserProfile = {
+  userId: 1,
+  nickName: '지식 헌터',
+  points: 2000,
+  applied: upcomingStudies.length,
+  in_progress: ongoingStudies.length,
+  completed: completedStudies.length,
+  profileImage: 'https://placehold.co/48x48',
+};
+
+export const avgStatsData: AvgStats = {
+  userId: 1,
+  userTodoCompletion: 100,
+  averageTodoCompletion: 65,
+  userStudyTime: 100,
+  averageStudyTime: 352,
+};
